Use findPackages approach is jest.config as well
